% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vowels.R
\name{vowel_strs}
\alias{vowel_strs}
\title{Extract vowels from a character vector}
\usage{
vowel_strs(x)
}
\arguments{
\item{x}{character vector.}
}
\description{
Iterate over a character vector, to extract vowels.
Return a named vector of the vowels found.
}
